+++ agent/src/main/java/eu/cqse/teamscale/jacoco/util/LoggingUtils.java
@@ -93,6 +93,11 @@ public class LoggingUtils {
 			}
 
 			@Override
+			public void info(String message) {
+				logger.info(message);
+			}
+
+			@Override
 			public void warn(String message) {
 				logger.warn(message);
 			}