+++ teamscale-gradle-plugin/src/main/kotlin/eu/cqse/ImpactedTestsExecutorTask.kt
@@ -219,6 +219,7 @@ fun Logger.wrapInILogger(): ILogger {
     val logger = this
     return object : ILogger {
         override fun debug(message: String) = logger.debug(message)
+        override fun info(message: String) = logger.info(message)
         override fun warn(message: String) = logger.warn(message)
         override fun warn(message: String, throwable: Throwable) = logger.warn(message, throwable)
         override fun error(throwable: Throwable) = logger.error("", throwable)